[bug] #87 : 단계 화면에서 홈버튼이 작동하지 않음
