+++ CleanFish/CleanFish/Assets/Images.xcassets/staticwave.imageset/Contents.json
@@ -2,16 +2,7 @@
   "images" : [
     {
       "filename" : "staticwave.png",
-      "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "idiom" : "universal",
-      "scale" : "2x"
-    },
-    {
-      "idiom" : "universal",
-      "scale" : "3x"
+      "idiom" : "universal"
     }
   ],
   "info" : {
